chore: #3 로그인 기능 2차 구현 22-11-15 20:45 (김민섭)

+++ controller/members.controller.js
@@ -61,6 +61,10 @@ class MembersController {
 
   deleteMember = async (req, res, next) => {
     try {
+      console.log(tokenInfo);
+      if (tokenInfo.message === "jwt expired") {
+        throw new Error("jwt expired");
+      }
       const { _id } = res.locals.members;
       await this.membersService.deleteMember(_id);
       res.status(200).send({ message: "회원탈퇴가 완료되었습니다" });
